fix edge cases with 'negative duration' (++ test)
- there were some edge cases observed where the duration
would be negative ... this happened from the started_at time
begin calculated directly from the 'created_at' time on the incident.
what really needed to happen is to go through the incident_updates,
created_at, and so on, and from all of that, pick the earliest time.
it seems this addresses cases where the incident was punched in
retroactively, but not all fields filled in. updated incident.rb.

- also saved one of those problem cases into the test fixtures.